+++ carver-pokedex3/src/App.js
@@ -1,9 +1,16 @@
 import Router from "./route/Router"
-
+import GlobalStateContext from "./contexts/GlobalStateContext";
+import { useEffect, useState } from 'react'
+import axios from 'axios'
+import GlobalPokemon from "./contexts/GlobalPokemon";
 
 const App = () => {
+  
+
   return (
-    <Router />
+    <GlobalPokemon>
+      <Router />
+    </GlobalPokemon>
   )
 }
