+++ carver-pokedex3/src/Pages/Main/Main.js
@@ -1,6 +1,6 @@
 import PokeCard from "../../components/PokeCard/PokeCard"
 import { MainContainer } from "./styled"
-import useRequestData from "../../components/useRequestData/useRequestData";
+
 
 
 const Main = () => {
