+++ carver-pokedex3/src/Pages/PokeDetails/PokeDetails.js
@@ -1,6 +1,6 @@
 import { ButtonsCard, CardContainer, PhotoCard } from "../../components/PokeCard/styled"
 
-const PokeDetails = () => {
+const PokeDetails = (props) => {
     return (
         <div>PokeDetails page</div>
     )
