+++ carver-pokedex3/src/contexts/GlobalPokemon.js
@@ -1,17 +1,45 @@
 import { useEffect, useState } from 'react'
 import axios from 'axios'
 import { url } from '../constants/urls'
+import GlobalStateContext from './GlobalStateContext'
 
 
-const GlobalPokemon = (props) =>{
+const GlobalPokemon = (props) => {
+  const [mostrarPokemon, setMostrarPokemon] = useState([])
 
-    const pokemon = useRequestData([], url)
-    return (
-        <GlobalContext.Provider value={{pokemon}}>
-          {props.children}
-        </GlobalContext.Provider>
-      )
+  useEffect(() => {
+    detalhesPokemon()
+  }, [])
+
+
+
+  const detalhesPokemon = () => {
+    const list = []
+    for (let i = 1; i <= 20; i++) {
+      axios.get(`https://pokeapi.co/api/v2/pokemon/${i}`)
+        .then((res) => {
+          list.push(res.data)
+          if (list.length === 20) {
+            setMostrarPokemon(list)
+          }
+        })
+        .catch((err) => {
+          console.log(err.response)
+        })
+    }
+  }
+ 
+  const data = {
+    mostrarPokemon,
+    setMostrarPokemon
+  }
+
+  return (
+    <GlobalStateContext.Provider value={data}>
+      {props.children}
+    </GlobalStateContext.Provider>
+  )
 }
 
-}
+
 export default GlobalPokemon
